cn-reels-pipeline/requirements.txt
manim==0.18.1
moviepy==1.0.3
requests==2.32.3
python-dotenv==1.0.1
pysrt==1.1.2
openai-whisper==20240930
pydub==0.25.1
Pillow>=10.0.0
numpy>=1.26.0
